Rebase fixes

+++ src/hotspot/share/utilities/growableArray.hpp
@@ -302,8 +302,8 @@ public:
   }
 
   void print_on(outputStream *st) const {
-    st->print("Growable Array " INTPTR_FORMAT, this);
-    st->print(": length %ld (_max %ld) { ", _len, _max);
+    st->print("Growable Array " INTPTR_FORMAT, (intptr_t) this);
+    st->print(": length %d (_max %d) { ", _len, _max);
     for (int i = 0; i < _len; i++) {
       st->print(INTPTR_FORMAT " ", *(intptr_t*)&(_data[i]));
     }
